Refactored proof validation and nonce handling by consolidating functionality into `proof` and `nonce` packages, introducing modular interfaces, and updating references accordingly.
